// interfaces/config.tsx
export interface Configuration {
  name: string;
  primaryColor: string;
  secondaryColor: string;
  backgroundColor: string;
  surfaceColor: string;
  tabBarPrimaryColor: string;
  tabBarSecondaryColor: string;
  textPrimaryColor: string;
  textSecondaryColor: string;
  buttonPrimaryColor: string;
  buttonSecondaryColor: string;
  borderColor: string;
}